requirements.txt
streamlit
pandas==1.5.3
openai-whisper
ffmpeg-python==0.2.0
textblob==0.15.3
torch==2.0.1
numpy==1.24.0
scipy==1.9.3
unicodedata2
